Large Changes
Moved debt page to overview, replaced debt page with a calculator for monthly payments.
