+++ client/src/components/Nav.js
@@ -85,7 +85,7 @@ const MenuLinks = ({ isOpen, loading, handleLogout, isAuth }) => {
                 { isAuth ? <MenuItem to="/overview">Overview</MenuItem> : null }
                 { isAuth ? <MenuItem to="/budget">Budget</MenuItem> : null }
                 { isAuth ? <MenuItem to="/transactions">Transactions</MenuItem> : null }
-                { isAuth ? <MenuItem to="/debt">Debt Breakdown</MenuItem> : null}
+                { isAuth ? <MenuItem to="/debt">Debt Calculator</MenuItem> : null}
                 { isAuth ? (
                     <Button
                         isLoading={loading}
