+++ client/src/routes/Pages/Overview/Debt.js
@@ -1,68 +1,292 @@
-import React, {useState, useEffect} from "react";
-import { useSelector } from 'react-redux'
+import React, { useState, useEffect } from "react";
+import { useDispatch, useSelector } from 'react-redux'
+
+// create pie charts using ApexCharts
+import ReactApexChart from "react-apexcharts";
 import { 
     Box,
     Text,
     Stack,
+    List,
+    ListItem,
+    ListIcon,
+    Button,
+    useColorModeValue, 
     SimpleGrid, 
-    Divider,
-    useColorModeValue
+    Center
 } from '@chakra-ui/react'
 
-
 const Debt = () => {
-    const allTransactions = useSelector((state) => state.transactions.transactions)
-    // make an array called transactions that is initialized to state
-    const [transactions, setTransactions] = useState([])
-    const [debt, setDebt] = useState(0)
-    const [month1, setMonth1] = useState('January')
-    const [month2, setMonth2] = useState('January')
-    const [month3, setMonth3] = useState('January')
-    const [thisMonth, setThisMonth] = useState([])
-    const [prevMonth, setPrevMonth] = useState([])
-    const [prevPrevMonth, setPrevPrevMonth] = useState([])
+    // Show spending by Bank/Credit, and by Category and separate by past week, month, year, all time
 
+    // display the stats in a table
+    
+    const allTransactions = useSelector((state) => state.transactions.transactions)
+    const dispatch = useDispatch();
+
+    // create an array of all the transactions that are debts
+    const [debts, setDebts] = useState([])
     const now = new Date()
 
-    useEffect(() => {
-        // just use allTransactions for now, no need to filter
-        setTransactions(allTransactions)
-    }, [allTransactions])
+    // given 'week', 'month', 'year', 'all', return the transactions that are in that time period
+    const getTransactions = (timePeriod) => {
+        let transactions = []
+        const lastWeek = new Date(now.getFullYear(), now.getMonth(), now.getDate() - 7)
+        const lastMonth = new Date(now.getFullYear(), now.getMonth() - 1, now.getDate())
+        const lastYear = new Date(now.getFullYear() - 1, now.getMonth(), now.getDate())
+        switch (timePeriod) {
+            case 'week':
+                // set transactions to all transactions that are within the past week
+                for (let i = 0; i < allTransactions.length; i++) {
+                    // translate date in MM/DD/YYYY format to a date object
+                    let date = new Date(allTransactions[i].date)
+                    if (date > lastWeek) {
+                        transactions.push(allTransactions[i])
+                    }
+                }
+                // console.log("week" + transactions)
+                break;
+            case 'month':
+                for (let i = 0; i < allTransactions.length; i++) {
+                    let date = new Date(allTransactions[i].date)
+                    if (date > lastMonth) {
+                        transactions.push(allTransactions[i])
+                    }
+                }
+                // console.log("month" + transactions)
+                break;
+            case 'year':
+                for (let i = 0; i < allTransactions.length; i++) {
+                    let date = new Date(allTransactions[i].date)
+                    if (date > lastYear) {
+                        transactions.push(allTransactions[i])
+                    }
+                }
+                // console.log("year" + transactions)
+                break;
+            case 'all':
+                transactions = allTransactions
+                break;
+            default:
+                transactions = allTransactions
+                break;
+        }
+        return transactions
+    }
 
-    useEffect(() => {
+    // given an array of transactions, and either 'source' or 'category', return an array of objects with the name and amount of the source/category
+    const getStats = (transactions, type) => {
+        let stats = []
         let total = 0
-        thisMonth.forEach(transaction => {
-            total += transaction.amount
+        transactions.forEach(t => {
+            if (type === 'source') {
+                let source = stats.find(s => s.name === t.source)
+                if (source) {
+                    source.amount += t.amount
+                } else {
+                    stats.push({name: t.source, amount: t.amount})
+                }
+            } else if (type === 'category') {
+                let category = stats.find(s => s.name === t.category)
+                if (category) {
+                    category.amount += t.amount
+                } else {
+                    stats.push({name: t.category, amount: t.amount})
+                }
+            }
+            total += t.amount
         })
-        setDebt(total)
-    }, [thisMonth])
+        return {stats, total}
+    }
+
+    const getPercentages = (stats, total) => {
+        let percentages = []
+        stats.forEach(s => {
+            percentages.push({name: s.name, percentage: s.amount / total})
+        })
+        return percentages
+    }
+
+    const getColors = (percentages) => {
+        let colors = []
+        percentages.forEach(p => {
+            let color = ''
+            if (p.percentage < 0.5) {
+                color = `rgb(${255 * (1 - p.percentage * 2)}, 255, 0)`
+            } else {
+                color = `rgb(255, ${255 * (1 - (p.percentage - 0.5) * 2)}, 0)`
+            }
+            colors.push({name: p.name,
+                color: color})
+        })
+        return colors
+    }
+
+    // Display the stats in a table
+    const [timePeriod, setTimePeriod] = useState('all')
+    const [sourceStats, setSourceStats] = useState([])
+    const [categoryStats, setCategoryStats] = useState([])
+    const [sourcePercentages, setSourcePercentages] = useState([])
+    const [categoryPercentages, setCategoryPercentages] = useState([])
+    const [sourceColors, setSourceColors] = useState([])
+    const [categoryColors, setCategoryColors] = useState([])
+    const [sourceTotal, setSourceTotal] = useState(0)
+    const [categoryTotal, setCategoryTotal] = useState(0)
+    
+    useEffect(() => {
+        const transactions = getTransactions(timePeriod)
+        const sourceStats = getStats(transactions, 'source')
+        const categoryStats = getStats(transactions, 'category')
+        const sourcePercentages = getPercentages(sourceStats.stats, sourceStats.total)
+        const categoryPercentages = getPercentages(categoryStats.stats, categoryStats.total)
+        const sourceColors = getColors(sourcePercentages)
+        const categoryColors = getColors(categoryPercentages)
+        setSourceStats(sourceStats.stats)
+        setCategoryStats(categoryStats.stats)
+        setSourcePercentages(sourcePercentages)
+        setCategoryPercentages(categoryPercentages)
+        setSourceColors(sourceColors)
+        setCategoryColors(categoryColors)
+        setSourceTotal(sourceStats.total)
+        setCategoryTotal(categoryStats.total)
+    }
+
+    , [timePeriod])
+
+    // create the pie charts
+    const [sourceSeries, setSourceSeries] = useState([])
+    const [categorySeries, setCategorySeries] = useState([])
+    const [sourceLabels, setSourceLabels] = useState([])
+    const [categoryLabels, setCategoryLabels] = useState([])
 
     useEffect(() => {
-        const d = new Date()
-        const months = ["January", "February", "March", "April", "May", "June","July","August","September","October","November","December"]
-        setMonth1(months[d.getMonth()])
-        setMonth2(months[d.getMonth() - 1])
-        setMonth3(months[d.getMonth() - 2])
-    }, [])
+        const sourceSeries = []
+        const sourceLabels = []
+        sourcePercentages.forEach(s => {
+            sourceSeries.push(s.percentage)
+            sourceLabels.push(s.name)
+        })
+        const categorySeries = []
+        const categoryLabels = []
+        categoryPercentages.forEach(c => {
+            categorySeries.push(c.percentage)
+            categoryLabels.push(c.name)
+        })
+        setSourceSeries(sourceSeries)
+        setCategorySeries(categorySeries)
+        setSourceLabels(sourceLabels)
+        setCategoryLabels(categoryLabels)
+    }, [sourcePercentages, categoryPercentages])
 
-    // thisMonth should be an array of transactions that are in the current month and year
-    useEffect(() => {
-        const thisMonth = transactions.filter(transaction => {
-            const d = new Date(transaction.date)
-            return d.getMonth() === now.getMonth() && d.getFullYear() === now.getFullYear()
-        })
-        setThisMonth(thisMonth)
-        const prevMonth = transactions.filter(transaction => {
-            const d = new Date(transaction.date)
-            return d.getMonth() === now.getMonth() - 1 && d.getFullYear() === now.getFullYear()
-        })
-        setPrevMonth(prevMonth)
-        const prevPrevMonth = transactions.filter(transaction => {
-            const d = new Date(transaction.date)
-            return d.getMonth() === now.getMonth() - 2 && d.getFullYear() === now.getFullYear()
-        })
-        setPrevPrevMonth(prevPrevMonth)
-    }, [transactions])
+    const sourceOptions = {
+        chart: {
+          type: "donut",
+        },
+        labels: sourceLabels,
+        plotOptions: {
+          pie: {
+            donut: {
+              labels: {
+                show: true,
+                name: {
+                  show: true,
+                  fontSize: "22px",
+                  fontFamily: "Helvetica, Arial, sans-serif",
+                  fontWeight: 600,
+                  color: undefined,
+                  offsetY: -10
+                },
+                value: {
+                  show: true,
+                  fontSize: "16px",
+                  fontFamily: "Helvetica, Arial, sans-serif",
+                  fontWeight: 400,
+                  color: undefined,
+                  offsetY: 16,
+                  formatter: function (val) {
+                    return "$" + val;
+                  }
+                },
+                total: {
+                  show: true,
+                  label: "Total",
+                  color: "#373d3f",
+                  formatter: function (w) {
+                    return "$" + sourceTotal;
+                  }
+                }
+              }
+            }
+          }
+        },
+        responsive: [{
+          breakpoint: 480,
+          options: {
+            chart: {
+              width: 200
+            },
+            legend: {
+              position: 'bottom'
+            }
+          }
+        }]
+      };
+      
+      const categoryOptions = {
+        chart: {
+          type: "donut",
+        },
+        labels: categoryLabels,
+        plotOptions: {
+            pie: {
+                donut: {
+                    labels: {
+                        show: true,
+                            name: {
+                                show: true,
+                                fontSize: "22px",
+                                fontFamily: "Helvetica, Arial, sans-serif",
+                                fontWeight: 600,
+                                color: undefined,
+                                offsetY: -10
+                                },
+                            value: {
+                                                    show: true,
+                                fontSize: "16px",
+                                fontFamily: "Helvetica, Arial, sans-serif",
+                                fontWeight: 400,
+                                color: undefined,
+                                offsetY: 16,
+                                formatter: function (val) {
+                                    return "$" + (categoryTotal * val).toFixed(2);
+                                }
+                            },
+                            total: {
+                                show: true,
+                                label: "Total",
+                                color: undefined,
+                                formatter: function (w) {
+                                    // truncate decimal values to 2 digits
+                                    return "$" + categoryTotal.toFixed(2);
+                                }
+                            }
+                        }
+                    }
+                }
+            },
+        responsive: [{
+            breakpoint: 480,
+            options: {
+                chart: {
+                    width: 200
+                },
+                legend: {
+                    position: 'bottom'
+                }
+            }
+        }]
+        };
+
 
     return (
         <Box
@@ -91,44 +315,53 @@ const Debt = () => {
                         </Stack>
                     </Stack>
                 </SimpleGrid>
-                <Box bg={useColorModeValue('gray.50', 'gray.900')} px={6} py={10}>
-                    {/* Display the current month and how much debt you have in that month */}
-                    <Stack spacing={0} align={'center'} justify={'center'}>
-                        <Text
-                            color={useColorModeValue('gray.800', 'white')}
-                            fontWeight={600}
-                            fontSize={'2xl'}>
-                            {month1}
-                        </Text>
-                        <Text fontSize={'4xl'} fontWeight={800} color={'black'}>
-                            ${debt}
-                        </Text>
-                    </Stack>
-                    {/* Display the previous month and how much debt you had in that month */}
-                    <Stack spacing={0} align={'center'} justify={'center'}>
-                        <Text
-                            color={useColorModeValue('gray.800', 'white')}
-                            fontWeight={600}
-                            fontSize={'2xl'}>
-                            {month2}
-                        </Text>
-                        <Text fontSize={'4xl'} fontWeight={800} color={'black'}>
-                            ${prevMonth.reduce((acc, transaction) => acc + transaction.amount, 0)}
-                        </Text>
-                    </Stack>
-                    {/* Display the month before that and how much debt you had in that month */}
-                    <Stack spacing={0} align={'center'} justify={'center'}>
-                        <Text
-                            color={useColorModeValue('gray.800', 'white')}
-                            fontWeight={600}
-                            fontSize={'2xl'}>
-                            {month3}
-                        </Text>
-                        <Text fontSize={'4xl'} fontWeight={800} color={'black'}>
-                            ${prevPrevMonth.reduce((acc, transaction) => acc + transaction.amount, 0)}
-                        </Text>
-                    </Stack>
-                </Box>
+                    <div style={{backgroundColor: '#f5f5f5', padding: '20px'}}>
+                    <div>
+                        <div style={{display: 'flex', justifyContent: 'center', backgroundColor: '#e0e0e0', padding: '10px'}}>
+                            <button onClick={() => setTimePeriod('week')}>Week</button>
+                            <div style={{width: '10px'}}></div>
+                            <button onClick={() => setTimePeriod('month')}>Month</button>
+                            <div style={{width: '10px'}}></div>
+                            <button onClick={() => setTimePeriod('year')}>Year</button>
+                            <div style={{width: '10px'}}></div>
+                            <button onClick={() => setTimePeriod('all')}>All</button>
+                        </div>
+                    </div>
+                    <div>
+                        <div style={{display: 'flex', justifyContent: 'center'}}>
+                            <ReactApexChart 
+                                options={categoryOptions} 
+                                series={categorySeries} 
+                                type="donut" 
+                                width={500}
+                            />
+                        </div>
+                        {/* <table>
+                            <thead>
+                                <tr>
+                                    <th>Category</th>
+                                    <div style={{width: '10px'}}></div>
+                                    <th>Amount</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                {categoryStats ? categoryStats.map(c => {
+                                    return (
+                                        <tr>
+                                            <td>{c.name ? c.name : 'Uncategorized'}</td>
+                                            <div style={{width: '10px'}}></div>
+                                            <td>{c.amount ? c.amount : 0}</td>
+                                        </tr>
+                                    )
+                                }) : 
+                                <tr>
+                                    <td>No data</td>
+                                </tr>
+                                }
+                            </tbody>
+                        </table> */}
+                    </div>
+                </div>
             </Box>
         </Box>
     )
